stage2/04-pirate-radio/files/usr/local/lib/radio-settings/requirements.txt
Django==3.2.6
django-apscheduler==0.6.0
feedparser==6.0.8
python-dateutil==2.8.2
requests==2.26.0
